requirements.txt
numpy>=1.18.2
torchvision 
tqdm
cudatoolkit >= 10.2
pytorch
wandb >= 0.10.23